Fix: moved type declarations to types file
