--- lib/types.d.ts
@@ -1,5 +1,0 @@
-export type IError = {
-    message: string;
-    stack: string;
-    code: any;
-} | null;
